feat(jobserver): Introduce BinaryDAO and MetaDataDAO
This will allow future storage implementations to be more
flexible. Binary and MetaData storage can be implemented independently
on different storages and combined later.

The goal is at some point refactor and remove JobDAO and make storage
more configurable.
Initial idea was to split JobDAO itself, but we ran into several
problems with this approach:
- Naming of new functions, e.g. saveBinary in MetaDataDAO and BinaryDAO
(name is already taken)
- Creating new name, e.g. saveBinaryInfo made it obvious that existing
DAOs have a lot of private functions (sometimes with same name, but
different return type). It required a lot of refactoring in existing SQL and
Cassandra DAO. The idea is to make it gracefully later.
